requirements.txt
pyqt6>=6.4.2
pyyaml>=6.0
pandas>=2.0.3